% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SQLite.R, R/create_fetchNASIS.R
\name{create_SQLite}
\alias{create_SQLite}
\alias{create_fetchNASIS_pedons}
\alias{create_fetchNASIS_components}
\title{Create a SQLite Database}
\usage{
create_SQLite(
  output_path = tempfile(fileext = ".sqlite"),
  tables = NULL,
  SS = FALSE
)

create_fetchNASIS_pedons(
  output_path = tempfile(fileext = ".sqlite"),
  tables = c("site", "pedon", "transect", "metadata", "lookup"),
  SS = FALSE
)

create_fetchNASIS_components(
  output_path = tempfile(fileext = ".sqlite"),
  tables = c("datamapunit", "component", "metadata", "lookup"),
  SS = FALSE
)
}
\arguments{
\item{output_path}{Output SQLite file}

\item{tables}{Passed to `soilDB::get_NASIS_table_name_by_purpose(purpose = ...)` with and without `SS = TRUE`}

\item{SS}{Use Selected Set (`..._View_1`)tables?}
}
\value{
SQLite file written to `output_path`
}
\description{
Create a SQLite Database
}
