% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDBdata-package.R
\docType{package}
\name{soilDBdata-package}
\alias{soilDBdata}
\alias{soilDBdata-package}
\title{soilDBdata: Create static data snapshots for soilDB unit tests and comparisons}
\description{
The goal of soilDBdata is to provide standard methods for building and provisioning test data sets for the soilDB R package. These datasets are stored as SQLite, JSON, CSV in the inst/extdata folder and are buildable by standard methods. Building datasets from scratch may require access to a local instance of a NASIS database with relevant tables populated.
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
