% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilDBdata-package.R
\docType{package}
\name{soilDBdata-package}
\alias{soilDBdata}
\alias{soilDBdata-package}
\title{soilDBdata: Create static data snapshots for soilDB unit tests and comparisons}
\description{
The goal of soilDBdata is to provide standard methods for building test data sets for the soilDB package.
}
\author{
\strong{Maintainer}: Andrew Brown \email{andrew.g.brown@usda.gov}

}
\keyword{internal}
